+++ pyproject.toml
@@ -41,7 +41,7 @@ path = "hatch_custom_hook.py"
 [tool.hatch.build.hooks.custom.copy_version_py]
 destinations = [
   "src/deadline/nuke_adaptor",
-  "src/deadline/nuke",
+  "src/deadline/nuke_util",
   "src/deadline/nuke_submitter",
 ]
 
@@ -88,7 +88,7 @@ xfail_strict = true
 addopts = [
     "--durations=5",
     "--cov=src/deadline/nuke_adaptor",
-    "--cov=src/deadline/nuke",
+    "--cov=src/deadline/nuke_util",
     "--cov=src/deadline/nuke_submitter",
     "--color=yes",
     "--cov-report=html:build/coverage",
@@ -107,7 +107,7 @@ filterwarnings = [
 ]
 
 [tool.coverage.run]
-source_pkgs = [ "deadline/nuke_adaptor", "deadline/nuke", "deadline/nuke_submitter" ]
+source_pkgs = [ "deadline/nuke_adaptor", "deadline/nuke_util", "deadline/nuke_submitter" ]
 branch = true
 parallel = true
 omit = [
