+++ src/deadline/nuke_adaptor/NukeClient/nuke_client.py
@@ -28,9 +28,10 @@ except ImportError:
 
 try:
     import nuke
-    from deadline.nuke import ocio_util
 except ImportError:  # pragma: no cover
     raise OSError("Could not find the Nuke module. Are you running this inside of Nuke?")
+
+from deadline.nuke_util import ocio_util
 
 
 class NukeClient(_HTTPClientInterface):
