+++ test/unit/deadline_submitter_for_nuke/test_assets.py
@@ -36,13 +36,13 @@ def _activated_reading_write_node_knobs(knob_name: str):
     "deadline.nuke_submitter.assets.get_node_filenames",
     return_value=["/one/asset.png", "/two/asset.png"],
 )
-@patch("deadline.nuke.ocio_util.is_custom_ocio_config_enabled", return_value=False)
+@patch("deadline.nuke_util.ocio_util.is_custom_ocio_config_enabled", return_value=False)
 @patch(
-    "deadline.nuke.ocio_util.get_custom_ocio_config_path",
+    "deadline.nuke_util.ocio_util.get_custom_ocio_config_path",
     return_value="/this/ocio_configs/config.ocio",
 )
 @patch(
-    "deadline.nuke.ocio_util.get_ocio_config_absolute_search_paths",
+    "deadline.nuke_util.ocio_util.get_ocio_config_absolute_search_paths",
     return_value=["/this/ocio_configs/luts"],
 )
 def test_get_scene_asset_references(
